# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Naratama Library API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the Naratama Library backend service."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.naratama.runsha.dev",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/rooms": {
      "get": {
        "tags": ["Rooms"],
        "summary": "List rooms",
        "parameters": [
          {
            "name": "available",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RoomType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rooms",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Room"
                          }
                        }
                      },
                      "required": ["count", "data"]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch rooms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/availability/{roomId}": {
      "parameters": [
        {
          "name": "roomId",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      ],
      "get": {
        "tags": ["Rooms"],
        "summary": "Check room availability",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability information",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "available": {
                          "type": "boolean"
                        },
                        "existingBookings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RoomBookingSlot"
                          }
                        }
                      },
                      "required": ["available", "existingBookings"]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to check availability",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/bookings": {
      "get": {
        "tags": ["Rooms"],
        "summary": "List room bookings",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "roomId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/BookingStatus"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged room bookings",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "page": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "totalPages": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RoomBookingDetailed"
                          }
                        }
                      },
                      "required": [
                        "count",
                        "page",
                        "total",
                        "totalPages",
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch room bookings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Rooms"],
        "summary": "Create room booking",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomBookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room booking created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoomBookingDetailed"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation or scheduling error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Staff or admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/bookings/{id}/status": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      ],
      "put": {
        "tags": ["Rooms"],
        "summary": "Update booking status",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomBookingStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking status updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoomBookingDetailed"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Staff or admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/bookings/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      ],
      "delete": {
        "tags": ["Rooms"],
        "summary": "Cancel booking",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Booking cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Staff or admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to cancel booking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie issued after successful login"
      }
    },
    "schemas": {
      "BaseSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["success"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": ["string", "object"],
            "nullable": true
          }
        },
        "required": ["success", "message"]
      },
      "ObjectId": {
        "type": "string",
        "pattern": "^[a-fA-F0-9]{24}$",
        "example": "64b7d2c6f2a5c2b1a0d4e5f6"
      },
...
      "BookingStatus": {
        "type": "string",
        "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED"]
      },
      "PaymentStatus": {
        "type": "string",
        "enum": ["UNPAID", "PAID", "REFUNDED"]
      },
      "RoomType": {
        "type": "string",
        "enum": ["SMALL_DISCUSSION", "LARGE_MEETING"]
      },
...
      "Membership": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "membershipNumber": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserSummary": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "membership": {
            "$ref": "#/components/schemas/Membership"
          }
        }
      },
...
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "roomNumber": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RoomType"
          },
          "capacity": {
            "type": "integer"
          },
          "hourlyRate": {
            "type": "number"
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isAvailable": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoomSummary": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "roomNumber": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RoomType"
          },
          "hourlyRate": {
            "type": "number"
          }
        }
      },
      "RoomBooking": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "roomId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "bookingDate": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "number"
          },
          "totalCost": {
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "purpose": {
            "type": "string",
            "nullable": true
          },
          "specialRequests": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoomBookingDetailed": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RoomBooking"
          },
          {
            "type": "object",
            "properties": {
              "room": {
                "$ref": "#/components/schemas/RoomSummary"
              },
              "user": {
                "$ref": "#/components/schemas/UserSummary"
              }
            }
          }
        ]
      },
      "RoomBookingSlot": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "number"
          }
        },
        "required": ["startTime", "endTime", "duration"]
      },
      "CreateRoomBookingRequest": {
        "type": "object",
        "properties": {
          "bookingDate": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "roomId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "purpose": {
            "type": "string"
          },
          "specialRequests": {
            "type": "string"
          }
        },
        "required": ["bookingDate", "startTime", "endTime", "roomId", "userId"]
      },
      "UpdateRoomBookingStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatus"
          }
        }
      }
    }
  }
}